+++ gmr/race_engine.py
@@ -1,0 +1,2683 @@
+# gmr/race_engine.py
+# Core race engine helpers (stage flow, AI stage choices, pit decisions)
+
+import random
+
+from gmr.constants import (
+    CHAMPIONSHIP_ACTIVE,
+    CONSTRUCTOR_SHARE,
+    POINTS_TABLE,
+    WEATHER_WET_CHANCE,
+    get_reliability_mult,
+    get_crash_mult,
+    TEST_DRIVERS_ENABLED,
+    get_prize_for_race_and_pos,
+)
+from gmr.data import drivers, tracks, constructors, engines, chassis_list
+from gmr.world_logic import driver_enters_event, get_car_speed_for_track, calculate_car_speed
+from gmr.careers import (
+    update_fame_after_race,
+    update_driver_progress,
+    grant_participation_xp_for_dnfs,
+    tick_driver_contract_after_race_end,
+)
+from gmr.story import maybe_trigger_demo_finale
+
+STAGE_LABELS = [
+    "Stage 1/3 — Opening Phase",
+    "Stage 2/3 — Mid-Race",
+    "Stage 3/3 — Final Push",
+]
+
+
+def per_stage_chance(total_chance, stages=3):
+    """
+    Convert a total event probability into an approximate per-stage chance.
+    Ensures the overall probability stays similar across multiple rolls.
+    """
+    total_chance = max(0.0, min(total_chance, 0.99))
+    if stages <= 1:
+        return total_chance
+    return 1.0 - (1.0 - total_chance) ** (1.0 / stages)
+
+
+def estimate_stage_wear(stage_index, race_length_factor, is_hot):
+    """
+    Rough UI-only estimates for tyre wear and engine temp by stage.
+    Returns (tyre_pct, engine_temp_pct) as 0–100.
+    """
+    base_tyre = 25 + (race_length_factor - 1.0) * 15
+    base_engine = 30 + (race_length_factor - 1.0) * 10
+
+    if is_hot:
+        base_engine += 15
+        base_tyre += 5
+
+    tyre = min(98, max(5, base_tyre * (stage_index + 1)))
+    engine = min(98, max(5, base_engine * (stage_index + 1)))
+    return int(tyre), int(engine)
+
+
+def generate_stage_blurb(stage_label, is_hot, is_wet, long_race, pace_note=None, incidents=None):
+    """
+    Procedural stage blurb including other-driver incidents and position changes.
+    """
+    segments = []
+
+    weather_openers = [
+        "Heat haze shimmers over the circuit.",
+        "The sun bakes the tarmac into a shimmering ribbon.",
+        "The crowd waves handkerchiefs in the glare of the afternoon heat.",
+        "Spray hangs in the air as grip comes and goes.",
+        "The rain eases, leaving a slippery sheen and scattered puddles.",
+        "The wind carries a fine mist across the braking zones.",
+        "The field settles into a steady rhythm.",
+        "Engines roar in unison as the pack strings out.",
+        "Marshals watch intently as the race finds its pace.",
+    ]
+
+    if is_hot:
+        segments.append(random.choice(weather_openers[:3]))
+    elif is_wet:
+        segments.append(random.choice(weather_openers[3:6]))
+    else:
+        segments.append(random.choice(weather_openers[6:]))
+
+    trackside = [
+        "A thin line of rubber darkens the racing groove.",
+        "Dust from the infield drifts onto the edge of the circuit.",
+        "Pit boards flash along the pit wall as crews take notes.",
+        "A brief lull lets drivers find a smoother line.",
+        "Spectators rise as two cars dispute the same corner.",
+        "The grandstands ripple with cheers as someone attacks into the braking zone.",
+    ]
+    if random.random() < 0.65:
+        segments.append(random.choice(trackside))
+
+    # Other-driver failures (named incidents)
+    if incidents:
+        for incident in incidents[:2]:
+            segments.append(incident)
+
+    # Other-driver failures (generic flavour)
+    failure_lines = [
+        "A plume of smoke signals trouble for a mid-pack runner.",
+        "A car crawls into the pits with a faltering engine.",
+        "Marshals wave yellow flags after a car stops off-line.",
+        "A privateer limps back with a misfire.",
+        "The crowd spots a stricken car at the far hairpin.",
+    ]
+    if random.random() < 0.60 and not incidents:
+        count = random.choice([1, 1, 2])
+        segments.append(
+            f"Around the circuit, {count} car{'s' if count != 1 else ''} slow or retire with trouble."
+        )
+        if random.random() < 0.55:
+            segments.append(random.choice(failure_lines))
+
+    # Overtake flavour
+    overtake_lines = [
+        "Your driver gets a better exit and draws alongside.",
+        "A brave move into the corner pays off.",
+        "A rival hesitates on the brakes and loses momentum.",
+        "A lapped car forces a momentary hesitation.",
+        "The slipstream tug proves decisive down the straight.",
+        "Your driver holds firm through the chicane.",
+    ]
+    if pace_note:
+        segments.append(pace_note)
+    elif random.random() < 0.75:
+        delta = random.choice([-3, -2, -1, 1, 2, 3])
+        if delta > 0:
+            segments.append(random.choice(overtake_lines))
+            segments.append(f"Your driver picks off {delta} rival{'s' if delta != 1 else ''} in traffic.")
+        else:
+            segments.append(random.choice(overtake_lines))
+            segments.append(f"A rival slips past — you drop {abs(delta)} place{'s' if abs(delta) != 1 else ''}.")
+
+    long_race_lines = [
+        "In this long race, pacing is already proving decisive.",
+        "The long distance begins to expose the fragile cars.",
+        "A steady hand seems more valuable with every passing kilometre.",
+    ]
+    if long_race and random.random() < 0.8:
+        segments.append(random.choice(long_race_lines))
+
+    tactical_lines = [
+        "The crew signals to keep it tidy through the rough section.",
+        "The pit wall urges patience as gaps fluctuate.",
+        "A brief lull lets your driver reset their rhythm.",
+        "Radio silence means the pit board is the only guide.",
+    ]
+    if random.random() < 0.6:
+        segments.append(random.choice(tactical_lines))
+
+    # Ensure multiple sentences for richer flavour
+    random.shuffle(segments)
+    return " ".join(segments[:min(len(segments), 5)])
+
+
+def choose_stage_strategy(stage_label, is_wet, is_hot):
+    """
+    Player choice for a race stage. Returns multipliers for performance, engine risk, crash risk.
+    """
+    print(f"\n=== Race Update: {stage_label} ===")
+    if is_wet:
+        print("Conditions are still tricky in the damp.")
+    elif is_hot:
+        print("Engines are running hot in the heat.")
+    else:
+        print("Track conditions are steady.")
+
+    print("Choose your approach for this phase:")
+    print("1) Push hard (faster, higher risk)")
+    print("2) Balanced (steady pace)")
+    print("3) Conserve (safer, slower)")
+
+    choice = input("> ").strip()
+
+    if choice == "1":
+        return {
+            "label": "Push hard",
+            "performance_mult": 1.04,
+            "engine_fail_mult": 1.25,
+            "crash_mult": 1.20,
+        }
+    if choice == "3":
+        return {
+            "label": "Conserve",
+            "performance_mult": 0.97,
+            "engine_fail_mult": 0.80,
+            "crash_mult": 0.85,
+        }
+
+    return {
+        "label": "Balanced",
+        "performance_mult": 1.00,
+        "engine_fail_mult": 1.00,
+        "crash_mult": 1.00,
+    }
+
+
+def maybe_stage_event(stage_label, is_wet, is_hot):
+    """
+    Random race-stage events that may affect the player's race.
+    Returns None or a dict with multipliers.
+    """
+    if random.random() > 0.35:
+        return None
+
+    events = [
+        {
+            "text": "Clear track ahead lets you build rhythm.",
+            "performance_mult": 1.02,
+            "engine_fail_mult": 1.00,
+            "crash_mult": 1.00,
+        },
+        {
+            "text": "Traffic slows your pace through the twisty section.",
+            "performance_mult": 0.98,
+            "engine_fail_mult": 1.00,
+            "crash_mult": 1.00,
+        },
+        {
+            "text": "A brake pedal goes long for a moment.",
+            "performance_mult": 0.99,
+            "engine_fail_mult": 1.00,
+            "crash_mult": 1.10,
+        },
+        {
+            "text": "The engine starts to run a little rough.",
+            "performance_mult": 0.98,
+            "engine_fail_mult": 1.12,
+            "crash_mult": 1.00,
+        },
+        {
+            "text": "A light shower begins, making the braking zones uncertain.",
+            "performance_mult": 0.99,
+            "engine_fail_mult": 1.00,
+            "crash_mult": 1.15,
+            "requires_wet": True,
+        },
+        {
+            "text": "Trackside dust drifts onto the racing line.",
+            "performance_mult": 0.99,
+            "engine_fail_mult": 1.00,
+            "crash_mult": 1.08,
+        },
+        {
+            "text": "A tailwind on the main straight boosts top speed.",
+            "performance_mult": 1.02,
+            "engine_fail_mult": 1.02,
+            "crash_mult": 1.00,
+        },
+        {
+            "text": "A vibration hints at a tire issue.",
+            "performance_mult": 1.00,
+            "engine_fail_mult": 1.00,
+            "crash_mult": 1.00,
+            "pit_issue": True,
+            "no_pit_performance_mult": 0.96,
+            "no_pit_crash_mult": 1.25,
+        },
+        {
+            "text": "A sudden tyre failure sends the driver scrambling to keep control.",
+            "performance_mult": 0.96,
+            "engine_fail_mult": 1.00,
+            "crash_mult": 1.20,
+            "forced_tyre_change": True,
+        },
+        {
+            "text": "Fuel delivery feels uneven under load.",
+            "performance_mult": 1.00,
+            "engine_fail_mult": 1.10,
+            "crash_mult": 1.00,
+            "pit_issue": True,
+            "no_pit_engine_mult": 1.20,
+            "no_pit_performance_mult": 0.97,
+        },
+        {
+            "text": "The car settles into a smooth rhythm.",
+            "performance_mult": 1.01,
+            "engine_fail_mult": 0.98,
+            "crash_mult": 0.98,
+        },
+        {
+            "text": "Oil temperature spikes briefly.",
+            "performance_mult": 0.99,
+            "engine_fail_mult": 1.12,
+            "crash_mult": 1.00,
+            "requires_hot": True,
+        },
+        {
+            "text": "A light drizzle makes the racing line tricky.",
+            "performance_mult": 0.99,
+            "engine_fail_mult": 1.00,
+            "crash_mult": 1.12,
+            "requires_wet": True,
+        },
+        {
+            "text": "A drying line appears and grip improves.",
+            "performance_mult": 1.02,
+            "engine_fail_mult": 1.00,
+            "crash_mult": 0.98,
+            "requires_wet": True,
+        },
+    ]
+
+    eligible = []
+    for ev in events:
+        if ev.get("requires_wet") and not is_wet:
+            continue
+        if ev.get("requires_hot") and not is_hot:
+            continue
+        eligible.append(ev)
+
+    if not eligible:
+        return None
+
+    event = random.choice(eligible)
+    return {
+        "text": event["text"],
+        "performance_mult": event.get("performance_mult", 1.0),
+        "engine_fail_mult": event.get("engine_fail_mult", 1.0),
+        "crash_mult": event.get("crash_mult", 1.0),
+        "pit_issue": event.get("pit_issue", False),
+        "forced_tyre_change": event.get("forced_tyre_change", False),
+        "no_pit_performance_mult": event.get("no_pit_performance_mult", 1.0),
+        "no_pit_engine_mult": event.get("no_pit_engine_mult", 1.0),
+        "no_pit_crash_mult": event.get("no_pit_crash_mult", 1.0),
+    }
+
+
+def maybe_stage_decision(stage_label, is_wet, is_hot):
+    """
+    Random decision prompts during a stage.
+    Returns None or a dict with two choices and multipliers.
+    """
+    if random.random() > 0.25:
+        return None
+
+    decisions = [
+        {
+            "prompt": "The driver reports fading brakes at the end of the straight.",
+            "choice_a": "Tell them to nurse the brakes",
+            "choice_b": "Tell them to keep pushing",
+            "a": {"performance_mult": 0.98, "engine_fail_mult": 0.98, "crash_mult": 0.90},
+            "b": {"performance_mult": 1.01, "engine_fail_mult": 1.02, "crash_mult": 1.15},
+        },
+        {
+            "prompt": "The pit board suggests conserving fuel. Do you comply?",
+            "choice_a": "Conserve fuel",
+            "choice_b": "Ignore and maintain pace",
+            "a": {"performance_mult": 0.98, "engine_fail_mult": 0.92, "crash_mult": 0.98},
+            "b": {"performance_mult": 1.01, "engine_fail_mult": 1.08, "crash_mult": 1.00},
+        },
+    ]
+
+    eligible = []
+    for dec in decisions:
+        if dec.get("requires_wet") and not is_wet:
+            continue
+        if dec.get("requires_hot") and not is_hot:
+            continue
+        eligible.append(dec)
+
+    if not eligible:
+        return None
+
+    return random.choice(eligible)
+
+
+def prompt_pit_decision(reason, prompt_text):
+    """
+    Ask player about a pit decision (1940s style: no radio).
+    """
+    print("\nPit decision:")
+    print(reason)
+    choice = input(prompt_text).strip().lower()
+    return choice == "y"
+
+
+def precompute_ai_stage_incidents(event_grid, state, track_profile, time, is_wet, is_hot, grid_risk_mult, race_length_factor):
+    """
+    Decide AI DNFs up front so stage updates can name actual incidents.
+    Returns dict: driver_name -> {"type": "engine"|"crash", "stage": label}.
+    """
+    incidents = {}
+
+    reliability_mult = get_reliability_mult(time)
+    crash_mult = get_crash_mult(time)
+
+    for d in event_grid:
+        if d == state.player_driver:
+            continue
+
+        car_speed, car_reliability = get_ai_car_stats(d.get("constructor"))
+        _ = car_speed  # unused here but kept for parity
+
+        mech = d.get("mechanical_sympathy", 5)
+        aggression = d.get("aggression", 5)
+        consistency = d.get("consistency", 5)
+        wet_skill = d.get("wet_skill", 5)
+
+        engine_fail_chance = (11 - car_reliability) * 0.02 * reliability_mult
+        engine_fail_chance *= (1 + (5 - mech) * 0.05)
+        engine_fail_chance *= track_profile.get("engine_danger", 1.0)
+        engine_fail_chance *= race_length_factor
+
+        if is_hot:
+            heat_intensity = track_profile.get("heat_intensity", 1.0)
+            engine_fail_chance *= heat_intensity
+
+        base_crash_chance = (11 - consistency) * 0.012
+        base_crash_chance *= (1 + (aggression - 5) * 0.05)
+        base_crash_chance *= (1 + (5 - mech) * 0.03)
+
+        crash_chance = base_crash_chance * crash_mult
+        crash_chance *= track_profile.get("crash_danger", 1.0)
+
+        if is_wet:
+            wet_factor = wet_skill / 10.0
+            rain_crash_mult = 1.40 - wet_factor * 0.30
+            crash_chance *= rain_crash_mult
+
+        sus = get_suspension_value_for_driver(state, d)
+        sus_importance = suspension_track_factor(track_profile)
+        crash_sus_mult = 1.08 - (sus - 5) * 0.02
+        crash_sus_mult = clamp(crash_sus_mult, 0.88, 1.15)
+        crash_sus_mult = 1.0 + (crash_sus_mult - 1.0) * sus_importance
+        crash_chance *= crash_sus_mult
+        crash_chance *= grid_risk_mult
+
+        ai_stage_mods = run_ai_race_stages(d, is_wet, is_hot, race_length_factor, car_reliability)
+        engine_fail_chance *= ai_stage_mods["engine_fail_mult"]
+        crash_chance *= ai_stage_mods["crash_mult"]
+
+        if random.random() < engine_fail_chance:
+            incidents[d.get("name")] = {
+                "type": "engine",
+                "stage": random.choice(STAGE_LABELS),
+            }
+            continue
+
+        if random.random() < crash_chance:
+            incidents[d.get("name")] = {
+                "type": "crash",
+                "stage": random.choice(STAGE_LABELS),
+            }
+
+    return incidents
+
+
+def precompute_ai_stage_overtakes(event_grid, quali_results):
+    """
+    Compare qualifying positions to predicted race finishing order.
+    Distribute position changes across stages as overtake announcements.
+    Returns dict: stage_label -> [("Smith", "Jones", 3)] meaning "Smith overtakes Jones for 3rd".
+    """
+    if not quali_results:
+        return {label: [] for label in STAGE_LABELS}
+
+    # Map driver names to their qualifying positions (1-indexed)
+    quali_positions = {}
+    for pos, (d, _) in enumerate(quali_results, start=1):
+        quali_positions[d.get("name")] = pos
+
+    # Predict race order using simple performance scoring (matches run_race logic)
+    predicted_race_order = []
+    for d in event_grid:
+        # Simple performance score (rough approximation)
+        score = d["pace"] * 1.2 + d["consistency"] * 0.4
+        predicted_race_order.append((d, score))
+
+    predicted_race_order.sort(key=lambda x: x[1], reverse=True)
+
+    # Map driver names to predicted race positions (1-indexed)
+    race_positions = {}
+    for pos, (d, _) in enumerate(predicted_race_order, start=1):
+        race_positions[d.get("name")] = pos
+
+    # Build reverse map: quali position -> driver name
+    quali_pos_to_name = {pos: name for name, pos in quali_positions.items()}
+    race_pos_to_name = {pos: name for name, pos in race_positions.items()}
+
+    # Distribute overtakes across stages (more in early/mid stages)
+    stage_overtakes = {label: [] for label in STAGE_LABELS}
+    stage_weights = [0.40, 0.40, 0.20]
+
+    # For each final race position, work backwards to see who was overtaken
+    all_overtakes = []
+    
+    for race_pos in sorted(race_positions.values()):
+        overtaker = race_pos_to_name.get(race_pos)
+        if not overtaker:
+            continue
+            
+        quali_pos_overtaker = quali_positions.get(overtaker)
+        if quali_pos_overtaker is None:
+            continue
+        
+        # If they gained positions (moved up)
+        if quali_pos_overtaker > race_pos:
+            # They overtook everyone between their quali pos and race pos
+            for pos_between in range(race_pos, quali_pos_overtaker):
+                # Who was in this position in quali?
+                overtaken = quali_pos_to_name.get(pos_between)
+                if overtaken and overtaken in race_positions:
+                    # Only record if the overtaken driver also finished
+                    all_overtakes.append({
+                        "overtaker": overtaker,
+                        "overtaken": overtaken,
+                        "new_position": race_pos
+                    })
+
+    # Assign each overtake to a stage
+    for overtake in all_overtakes:
+        # Pick a stage probabilistically
+        rand = random.random()
+        cumulative = 0.0
+        assigned_stage = STAGE_LABELS[-1]  # default to final stage
+
+        for idx, weight in enumerate(stage_weights):
+            cumulative += weight
+            if rand < cumulative:
+                assigned_stage = STAGE_LABELS[idx]
+                break
+
+        stage_overtakes[assigned_stage].append(
+            (overtake["overtaker"], overtake["overtaken"], overtake["new_position"])
+        )
+
+    return stage_overtakes
+
+
+def compute_stage_overtakes_from_results(quali_results, finishers):
+    """
+    Compute overtakes based on actual final results.
+    Uses a swap-based transform from quali order to final order.
+    Returns dict: stage_label -> [(overtaker, overtaken, new_pos), ...]
+    """
+    stage_overtakes = {label: [] for label in STAGE_LABELS}
+
+    if not quali_results or not finishers:
+        return stage_overtakes
+
+    finishers_names = [d.get("name") for d, _ in finishers]
+    quali_order = [d.get("name") for d, _ in quali_results if d.get("name") in finishers_names]
+    final_order = finishers_names[:]
+
+    if not quali_order or not final_order:
+        return stage_overtakes
+
+    current = quali_order[:]
+    swap_events = []
+
+    for target_index, target_name in enumerate(final_order):
+        if target_name not in current:
+            continue
+        current_index = current.index(target_name)
+        while current_index > target_index:
+            overtaker = current[current_index]
+            overtaken = current[current_index - 1]
+            current[current_index - 1], current[current_index] = current[current_index], current[current_index - 1]
+            current_index -= 1
+            new_pos = current_index + 1
+            swap_events.append((overtaker, overtaken, new_pos))
+
+    total = len(swap_events)
+    if total == 0:
+        return stage_overtakes
+
+    stage_weights = [0.40, 0.40, 0.20]
+    counts = [int(total * w) for w in stage_weights]
+    while sum(counts) < total:
+        for i in range(len(counts)):
+            if sum(counts) < total:
+                counts[i] += 1
+
+    idx = 0
+    for stage_idx, stage_label in enumerate(STAGE_LABELS):
+        take = counts[stage_idx]
+        for _ in range(take):
+            if idx >= total:
+                break
+            stage_overtakes[stage_label].append(swap_events[idx])
+            idx += 1
+
+    return stage_overtakes
+
+
+def consume_tyre_set(state, race_name, reason):
+    """
+    Consume one tyre set if available. Returns True if consumed.
+    """
+    if getattr(state, "tyre_sets", 0) <= 0:
+        state.news.append(
+            f"TYRES ({race_name}): No spare sets available — {reason} postponed."
+        )
+        return False
+
+    state.tyre_sets -= 1
+    state.news.append(
+        f"TYRES ({race_name}): Used 1 set ({reason}). Remaining: {state.tyre_sets}."
+    )
+    return True
+
+
+def run_player_race_stages(state, race_name, is_wet, is_hot, track_profile, stage_incidents=None, stage_overtakes=None, defer_output=False):
+    """
+    Run a simple 3-part race flow with player decisions and events.
+    Returns cumulative multipliers for performance and risks.
+    """
+    stage_defs = STAGE_LABELS
+    stage_output = []
+
+    perf_mult = 1.0
+    engine_mult = 1.0
+    crash_mult = 1.0
+
+    race_distance_km = track_profile.get("race_distance_km", 250.0)
+    long_race = race_distance_km >= 350.0
+
+    for idx, stage_label in enumerate(stage_defs):
+        tyre_pct, engine_temp_pct = estimate_stage_wear(idx, race_distance_km / 250.0, is_hot)
+        stage_events = []
+        stage_news = []
+        stage_blurb = None
+
+        # Collect ALL incidents for this stage
+        incidents = []
+        if stage_incidents:
+            for name, info in stage_incidents.items():
+                if info["stage"] == stage_label:
+                    if info["type"] == "engine":
+                        incidents.append(f"{name} slows with engine trouble")
+                    else:
+                        incidents.append(f"{name} spins into the barriers")
+
+        for incident in incidents:
+            stage_events.append(f"Incident: {incident}")
+
+        # Collect ALL overtakes for this stage
+        overtakes = []
+        if stage_overtakes and stage_label in stage_overtakes:
+            for overtaker, overtaken, new_pos in stage_overtakes[stage_label]:
+                # Add flavor to the overtake announcement
+                flavors = [
+                    f"{overtaker} slips past {overtaken} to grab P{new_pos}",
+                    f"{overtaker} outbrakes {overtaken} into P{new_pos}",
+                    f"{overtaker} powers past {overtaken} to take P{new_pos}",
+                    f"{overtaker} dives inside {overtaken} for P{new_pos}",
+                    f"{overtaker} overtakes {overtaken} to move into P{new_pos}",
+                    f"{overtaker} sweeps past {overtaken} to claim P{new_pos}",
+                ]
+                overtakes.append(random.choice(flavors))
+
+        choice = choose_stage_strategy(stage_label, is_wet, is_hot)
+        perf_mult *= choice["performance_mult"]
+        engine_mult *= choice["engine_fail_mult"]
+        crash_mult *= choice["crash_mult"]
+
+        stage_news.append(
+            f"RACE STRATEGY ({race_name}): {stage_label} — {choice['label']}."
+        )
+
+        event = maybe_stage_event(stage_label, is_wet, is_hot)
+        if event:
+            perf_mult *= event["performance_mult"]
+            engine_mult *= event["engine_fail_mult"]
+            crash_mult *= event["crash_mult"]
+            stage_events.append(event["text"])
+            stage_news.append(f"RACE EVENT ({race_name}): {event['text']}")
+
+            if event.get("forced_tyre_change"):
+                stage_events.append("The driver dives into the pits for a tyre change.")
+                if consume_tyre_set(state, race_name, "emergency tyre change"):
+                    perf_mult *= 0.96
+                    crash_mult *= 0.95
+                else:
+                    perf_mult *= 0.95
+                    crash_mult *= 1.15
+
+            if event.get("pit_issue"):
+                stage_events.append("The driver brings the car in to report the issue (no radio in this era).")
+                if prompt_pit_decision(
+                    "You can attempt a quick fix/tyre change, or send them straight back out.",
+                    "Attempt a quick fix? (y/n): "
+                ):
+                    if consume_tyre_set(state, race_name, "repair stop"):
+                        perf_mult *= 0.97
+                        engine_mult *= 0.90
+                        crash_mult *= 0.90
+                        perf_mult *= 1.01
+                        stage_news.append(f"PIT STOP ({race_name}): Team makes a quick repair stop.")
+                    else:
+                        perf_mult *= 0.985
+                        engine_mult *= event.get("no_pit_engine_mult", 1.0)
+                        crash_mult *= event.get("no_pit_crash_mult", 1.0)
+                        stage_news.append(f"PIT STOP ({race_name}): No tyres available; driver waved back out.")
+                else:
+                    perf_mult *= 0.985
+                    perf_mult *= event.get("no_pit_performance_mult", 1.0)
+                    engine_mult *= event.get("no_pit_engine_mult", 1.0)
+                    crash_mult *= event.get("no_pit_crash_mult", 1.0)
+                    stage_news.append(f"PIT STOP ({race_name}): Driver waved back out without repairs.")
+
+        if long_race and stage_label != stage_defs[0]:
+            if prompt_pit_decision(
+                "Long race. You can hang a pit board and call them in for fuel/tyres.",
+                "Signal for a stop? (y/n): "
+            ):
+                if consume_tyre_set(state, race_name, "fuel/tyre stop"):
+                    perf_mult *= 0.97
+                    perf_mult *= 1.01
+                    engine_mult *= 0.92
+                    crash_mult *= 0.92
+                    stage_news.append(f"PIT STOP ({race_name}): Fuel/tyre stop taken in a long race.")
+                else:
+                    perf_mult *= 0.98
+                    engine_mult *= 1.05
+                    crash_mult *= 1.05
+                    stage_news.append(f"PIT STOP ({race_name}): No tyres available; stop cancelled.")
+
+        decision = maybe_stage_decision(stage_label, is_wet, is_hot)
+        if decision:
+            print(f"\nDecision: {decision['prompt']}")
+            print(f"1) {decision['choice_a']}")
+            print(f"2) {decision['choice_b']}")
+            choice = input("> ").strip()
+            if choice == "1":
+                chosen = decision["a"]
+                stage_news.append(f"DECISION ({race_name}): {decision['choice_a']}.")
+            else:
+                chosen = decision["b"]
+                stage_news.append(f"DECISION ({race_name}): {decision['choice_b']}.")
+
+            perf_mult *= chosen.get("performance_mult", 1.0)
+            engine_mult *= chosen.get("engine_fail_mult", 1.0)
+            crash_mult *= chosen.get("crash_mult", 1.0)
+
+        # Stage pace note based on current multiplier
+        if perf_mult >= 1.03:
+            pace_note = "Your driver appears to be increasing the pace."
+        elif perf_mult <= 0.98:
+            pace_note = "Your driver looks to be losing ground."
+        else:
+            pace_note = "The pace looks steady for now."
+
+        stage_blurb = generate_stage_blurb(stage_label, is_hot, is_wet, long_race, pace_note=pace_note, incidents=[])
+
+        stage_output.append({
+            "label": stage_label,
+            "tyre_pct": tyre_pct,
+            "engine_temp_pct": engine_temp_pct,
+            "events": stage_events,
+            "blurb": stage_blurb,
+            "incidents": incidents,
+            "overtakes": overtakes,
+            "news": stage_news,
+        })
+
+        if not defer_output:
+            render_player_stage_output(state, race_name, [stage_output[-1]], stage_overtakes)
+
+    stage_mods = {
+        "performance_mult": perf_mult,
+        "engine_fail_mult": engine_mult,
+        "crash_mult": crash_mult,
+    }
+
+    return stage_mods, stage_output
+
+
+def render_player_stage_output(state, race_name, stage_output, stage_overtakes=None):
+    def highlight_player(text: str) -> str:
+        player = getattr(state, "player_driver", None)
+        if not player:
+            return text
+        name = player.get("name")
+        if not name:
+            return text
+        green = "\x1b[32m"
+        reset = "\x1b[0m"
+        return text.replace(name, f"{green}{name}{reset}")
+
+    for stage in stage_output:
+        stage_label = stage["label"]
+        tyre_pct = stage["tyre_pct"]
+        engine_temp_pct = stage["engine_temp_pct"]
+
+        print(f"\n[{stage_label}] Tyre wear estimate: {tyre_pct}% | Engine temp: {engine_temp_pct}%")
+
+        for entry in stage.get("news", []):
+            state.news.append(entry)
+
+        for event_text in stage.get("events", []):
+            print(event_text)
+
+        blurb = stage.get("blurb")
+        if blurb:
+            state.news.append(f"STAGE UPDATE ({race_name}): {blurb}")
+            print(highlight_player(blurb))
+
+        incidents = stage.get("incidents", [])
+        if incidents:
+            print(f"\nIncidents in {stage_label}:")
+            for incident in incidents:
+                print(f"  • {highlight_player(incident)}")
+                state.news.append(f"INCIDENT ({race_name}, {stage_label}): {incident}")
+
+        overtakes = stage.get("overtakes", [])
+        if stage_overtakes and stage_label in stage_overtakes:
+            overtakes = []
+            for overtaker, overtaken, new_pos in stage_overtakes[stage_label]:
+                flavors = [
+                    f"{overtaker} slips past {overtaken} to grab P{new_pos}",
+                    f"{overtaker} outbrakes {overtaken} into P{new_pos}",
+                    f"{overtaker} powers past {overtaken} to take P{new_pos}",
+                    f"{overtaker} dives inside {overtaken} for P{new_pos}",
+                    f"{overtaker} overtakes {overtaken} to move into P{new_pos}",
+                    f"{overtaker} sweeps past {overtaken} to claim P{new_pos}",
+                ]
+                overtakes.append(random.choice(flavors))
+
+        if overtakes:
+            print(f"\nPosition changes in {stage_label}:")
+            for overtake in overtakes:
+                print(f"  • {highlight_player(overtake)}")
+                state.news.append(f"OVERTAKE ({race_name}, {stage_label}): {overtake}")
+
+
+def render_stage_overtakes_only(state, race_name, stage_overtakes):
+    def highlight_player(text: str) -> str:
+        player = getattr(state, "player_driver", None)
+        if not player:
+            return text
+        name = player.get("name")
+        if not name:
+            return text
+        green = "\x1b[32m"
+        reset = "\x1b[0m"
+        return text.replace(name, f"{green}{name}{reset}")
+
+    if not stage_overtakes:
+        return
+
+    for stage_label in STAGE_LABELS:
+        overtakes = stage_overtakes.get(stage_label, [])
+        if not overtakes:
+            continue
+
+        print(f"\nPosition changes in {stage_label}:")
+        for overtaker, overtaken, new_pos in overtakes:
+            flavors = [
+                f"{overtaker} slips past {overtaken} to grab P{new_pos}",
+                f"{overtaker} outbrakes {overtaken} into P{new_pos}",
+                f"{overtaker} powers past {overtaken} to take P{new_pos}",
+                f"{overtaker} dives inside {overtaken} for P{new_pos}",
+                f"{overtaker} overtakes {overtaken} to move into P{new_pos}",
+                f"{overtaker} sweeps past {overtaken} to claim P{new_pos}",
+            ]
+            line = random.choice(flavors)
+            print(f"  • {highlight_player(line)}")
+            state.news.append(f"OVERTAKE ({race_name}, {stage_label}): {line}")
+
+
+def choose_ai_stage_strategy(stage_label, driver, is_wet, is_hot, race_length_factor, car_reliability):
+    """
+    AI stage choice for a race segment.
+    """
+    aggression = driver.get("aggression", 5)
+    consistency = driver.get("consistency", 5)
+    mech = driver.get("mechanical_sympathy", 5)
+
+    # Baseline tendency
+    push_bias = 0
+    conserve_bias = 0
+
+    if aggression >= 7:
+        push_bias += 1
+    if consistency >= 7:
+        push_bias += 1
+    if mech >= 7:
+        conserve_bias += 1
+    if car_reliability < 5:
+        conserve_bias += 2
+
+    if is_hot and mech <= 4:
+        conserve_bias += 1
+    if is_wet and consistency <= 4:
+        conserve_bias += 1
+
+    if race_length_factor >= 1.2 and stage_label != "Stage 3/3 — Final Push":
+        conserve_bias += 1
+
+    # Small randomness so AI isn't deterministic
+    roll = random.random()
+
+    if push_bias - conserve_bias >= 2 or (roll < 0.2 and push_bias > conserve_bias):
+        return "push"
+    if conserve_bias - push_bias >= 2 or (roll < 0.2 and conserve_bias > push_bias):
+        return "conserve"
+
+    return "balanced"
+
+
+def run_ai_race_stages(driver, is_wet, is_hot, race_length_factor, car_reliability):
+    """
+    Run 3-stage AI race flow and return combined multipliers.
+    """
+    stage_defs = STAGE_LABELS
+
+    perf_mult = 1.0
+    engine_mult = 1.0
+    crash_mult = 1.0
+
+    for stage_label in stage_defs:
+        choice = choose_ai_stage_strategy(
+            stage_label,
+            driver,
+            is_wet,
+            is_hot,
+            race_length_factor,
+            car_reliability,
+        )
+
+        if choice == "push":
+            perf_mult *= 1.03
+            engine_mult *= 1.15
+            crash_mult *= 1.12
+        elif choice == "conserve":
+            perf_mult *= 0.98
+            engine_mult *= 0.88
+            crash_mult *= 0.90
+
+    return {
+        "performance_mult": perf_mult,
+        "engine_fail_mult": engine_mult,
+        "crash_mult": crash_mult,
+    }
+
+
+def choose_ai_race_strategy(driver, constructor):
+    """
+    AI drivers choose strategy based on aggression and constructor.
+    Returns risk_mode string.
+    """
+    aggression = driver.get("aggression", 5)
+
+    # Base chance from aggression
+    if aggression >= 7:
+        attack_chance = 0.6
+        nurse_chance = 0.1
+    elif aggression <= 3:
+        attack_chance = 0.1
+        nurse_chance = 0.6
+    else:
+        attack_chance = 0.3
+        nurse_chance = 0.3
+
+    # Constructor influence
+    if constructor == "Enzoni":
+        attack_chance += 0.2  # More aggressive
+    elif constructor == "Independent":
+        nurse_chance += 0.1  # More conservative
+
+    # Normalize
+    attack_chance = min(attack_chance, 1.0)
+    nurse_chance = min(nurse_chance, 1.0)
+    neutral_chance = 1.0 - attack_chance - nurse_chance
+    neutral_chance = max(neutral_chance, 0.0)
+
+    rand = random.random()
+    if rand < attack_chance:
+        return "attack"
+    if rand < attack_chance + neutral_chance:
+        return "neutral"
+    return "nurse"
+
+
+def clamp(v, lo, hi):
+    return max(lo, min(hi, v))
+
+
+def pay_appearance_money(state, race_name):
+    t = tracks.get(race_name, {})
+    base = int(t.get("appearance_base", 0))
+    mult = float(t.get("appearance_prestige_mult", 0.0))
+
+    if base <= 0 and mult <= 0:
+        return 0
+
+    payout = int(base + state.prestige * mult)
+    payout = max(0, payout)
+
+    # Cap appearance money based on track prestige
+    if race_name in ["Bradley Fields", "Little Autodromo"]:
+        payout = min(payout, 50)
+    elif race_name in ["Ardennes Endurance GP", "Union Speedway"]:
+        payout = min(payout, 150)
+    else:
+        payout = min(payout, 100)
+
+    # Track appearance money separately from prize money
+    if not hasattr(state, "last_week_appearance_income"):
+        state.last_week_appearance_income = 0
+
+    state.money += payout
+    state.last_week_income += payout
+    state.last_week_appearance_income += payout
+
+    team_name = state.player_constructor or "Your team"
+    state.news.append(
+        f"Appearance money: {team_name} receive £{payout} from {race_name}'s organisers for taking the start."
+    )
+    return payout
+
+
+def get_suspension_value_for_driver(state, d):
+    """
+    Returns suspension score (1–10) for this driver's current car.
+    Player uses current_chassis. AI uses constructor chassis if available.
+    """
+    if d == state.player_driver and state.current_chassis:
+        return int(state.current_chassis.get("suspension", 5))
+
+    # AI: try to read from their works chassis
+    c = constructors.get(d.get("constructor", ""), {})
+    ch_id = c.get("chassis_id")
+    if ch_id:
+        ch = next((x for x in chassis_list if x.get("id") == ch_id), None)
+        if ch:
+            return int(ch.get("suspension", 5))
+
+    return 5
+
+
+def suspension_track_factor(track_profile):
+    return float(track_profile.get("suspension_importance", 1.0))
+
+
+def get_ai_car_stats(constructor_name):
+    c = constructors.get(constructor_name, {})
+
+    # Parts-based path FIRST (only if ids are defined)
+    eng_id = c.get("engine_id")
+    ch_id = c.get("chassis_id")
+    if eng_id and ch_id:
+        eng = next((e for e in engines if e["id"] == eng_id), None)
+        ch = next((x for x in chassis_list if x["id"] == ch_id), None)
+        if eng and ch:
+            speed = calculate_car_speed(eng, ch)
+            reliability = eng.get("reliability", 5)
+            return speed, reliability
+
+    # Legacy fallback (Independents stay here)
+    if "speed" in c and "reliability" in c:
+        return c["speed"], c["reliability"]
+
+    return 5, 5
+
+
+def build_event_grid(state, time, race_name, track_profile):
+    """
+    Returns a list of drivers who will take part in THIS event,
+    respecting per-team car limits.
+
+    IMPORTANT:
+    - "Independent" is NOT a real team; it's the open-entry pool.
+      So it should NOT be capped to 2 cars.
+    """
+
+    # Optional: let tracks define grid size (fallback to something sane)
+    grid_size = track_profile.get("grid_size", 12)
+
+    # How many cars each WORKS team can enter this year
+    def team_car_limit(team_name: str) -> int:
+        if team_name == "Enzoni":
+            return 3 if time.year >= 1950 else 2
+        if team_name == "Scuderia Valdieri":
+            return 2
+        if state.player_constructor and team_name == state.player_constructor:
+            return 1
+        return 999  # any future "real" team won't be accidentally capped
+
+    # Collect eligible drivers by "team"/pool
+    by_team = {}
+    for d in drivers:
+        if not driver_enters_event(d, race_name, track_profile, state):
+            continue
+
+        # Block Test drivers if debug toggle is off
+        if not TEST_DRIVERS_ENABLED and d.get("constructor") == "Test":
+            continue
+
+        # Player team: only the contracted player driver is allowed to represent it
+        if state.player_constructor and d.get("constructor") == state.player_constructor:
+            if d is not state.player_driver:
+                continue
+
+        team = d.get("constructor", "Independent")
+        by_team.setdefault(team, []).append(d)
+
+    final_grid = []
+
+    # 1) Add works teams and player (capped)
+    for team in ("Enzoni", "Scuderia Valdieri"):
+        if team in by_team:
+            team_drivers = by_team[team]
+
+            # best drivers get the seats
+            team_drivers.sort(
+                key=lambda x: (x.get("pace", 0) * 1.0 + x.get("consistency", 0) * 0.5),
+                reverse=True,
+            )
+
+            final_grid.extend(team_drivers[:team_car_limit(team)])
+
+    # Player
+    if state.player_constructor and state.player_constructor in by_team:
+        final_grid.extend(by_team[state.player_constructor][:1])
+
+    # 2) Fill the rest of the grid with Independents (open pool)
+    independents = by_team.get("Independent", [])
+
+    # IMPORTANT: don't always pick the same top guys
+    # Shuffle first, then lightly sort by ability so it still feels “real”
+    random.shuffle(independents)
+    independents.sort(
+        key=lambda x: (x.get("pace", 0) * 1.0 + x.get("consistency", 0) * 0.4),
+        reverse=True,
+    )
+
+    # Fill remaining slots up to grid_size
+    remaining = max(0, grid_size - len(final_grid))
+    final_grid.extend(independents[:remaining])
+
+    return final_grid
+
+
+def roll_race_weather(track_profile):
+    """
+    Decide race-day weather once, so you can show a forecast before the player
+    picks their risk level.
+    Returns (is_wet, is_hot).
+    """
+    track_wet_chance = track_profile.get("wet_chance", WEATHER_WET_CHANCE)
+    is_wet = random.random() < track_wet_chance
+
+    base_hot_chance = track_profile.get("base_hot_chance", 0.2)
+    is_hot = False
+    if not is_wet:
+        is_hot = random.random() < base_hot_chance
+
+    return is_wet, is_hot
+
+
+def record_race_result(state, time, season_week, race_name, is_wet, is_hot, finishers, retired):
+    """
+    Save one race to state.race_history and update state.driver_career totals.
+
+    finishers: list of (driver_dict, performance)
+    retired: list of (driver_dict, reason) where reason is "engine" or "crash" (or "unknown")
+    """
+
+    # Safety for old saves
+    if not hasattr(state, "race_history") or state.race_history is None:
+        state.race_history = []
+    if not hasattr(state, "driver_career") or state.driver_career is None:
+        state.driver_career = {}
+
+    entry = {
+        "year": time.year,
+        "week": season_week,
+        "race": race_name,
+        "wet": bool(is_wet),
+        "hot": bool(is_hot),
+        "finishers": [],
+        "dnfs": [],
+    }
+
+    # ---- finishers ----
+    for pos, (d, _perf) in enumerate(finishers, start=1):
+        pts = 0
+        if CHAMPIONSHIP_ACTIVE and (pos - 1) < len(POINTS_TABLE):
+            pts = POINTS_TABLE[pos - 1]
+
+        prize = get_prize_for_race_and_pos(race_name, pos - 1)
+
+        entry["finishers"].append({
+            "pos": pos,
+            "name": d.get("name"),
+            "constructor": d.get("constructor"),
+            "pts": pts,
+            "prize": prize,
+        })
+
+        name = d.get("name")
+        c = state.driver_career.get(name, {
+            "starts": 0,
+            "wins": 0,
+            "podiums": 0,
+            "dnfs": 0,
+            "engine_dnfs": 0,
+            "crash_dnfs": 0,
+            "points": 0,
+            "prize_money": 0,
+            "best_finish": None,
+        })
+
+        c["starts"] += 1
+        c["points"] += pts
+        c["prize_money"] += prize
+
+        # wins / podiums
+        if pos == 1:
+            c["wins"] += 1
+            c["podiums"] += 1
+        elif pos <= 3:
+            c["podiums"] += 1
+
+        # best finish
+        if c["best_finish"] is None or pos < c["best_finish"]:
+            c["best_finish"] = pos
+
+        state.driver_career[name] = c
+
+    # ---- DNFs ----
+    for d, reason in retired:
+        entry["dnfs"].append({
+            "name": d.get("name"),
+            "constructor": d.get("constructor"),
+            "reason": reason,
+        })
+
+        name = d.get("name")
+        c = state.driver_career.get(name, {
+            "starts": 0,
+            "wins": 0,
+            "podiums": 0,
+            "dnfs": 0,
+            "engine_dnfs": 0,
+            "crash_dnfs": 0,
+            "points": 0,
+            "prize_money": 0,
+            "best_finish": None,
+        })
+
+        c["starts"] += 1
+        c["dnfs"] += 1
+        if reason == "engine":
+            c["engine_dnfs"] += 1
+        if reason == "crash":
+            c["crash_dnfs"] += 1
+
+        state.driver_career[name] = c
+
+    state.race_history.append(entry)
+
+
+def add_engine_failure_explanation(
+    state,
+    d,
+    track_profile,
+    is_hot,
+    is_wet,
+    perspective="neutral",
+    breakdown=None,
+):
+    if breakdown is None:
+        breakdown = [("a technical failure", 1.0)]
+
+    top = breakdown[0][0] if breakdown else "a technical failure"
+
+    if perspective == "player":
+        state.news.append(
+            f"Post-race inspection suggests {top} was the decisive factor in the failure."
+        )
+    else:
+        state.news.append(
+            f"Paddock analysis points to {top} as the likely cause."
+        )
+
+
+def add_crash_explanation(state, d, track_profile, is_hot, is_wet, perspective="neutral", breakdown=None):
+    if breakdown is None:
+        breakdown = [("a simple driving error", 1.0)]
+
+    top = breakdown[0][0] if breakdown else "a simple driving error"
+
+    if perspective == "player":
+        state.news.append(
+            f"Race control cite {top} as the key factor in the accident."
+        )
+    else:
+        state.news.append(
+            f"Observers suggest {top} played a major role in the crash."
+        )
+
+
+def run_ai_only_race(state, race_name, time, season_week, track_profile):
+    """
+    AI-only race simulation for weeks where the player does not/cannot compete.
+
+    IMPORTANT:
+    - This version includes engine failures + crashes (DNFs), using the same core logic
+      style as run_race(), but without player wear/damage effects.
+    - Fame/XP update is applied to FINISHERS only (DNFs don't gain post-race fame).
+    """
+
+    # Roll conditions for flavour + crash modifiers
+    is_wet, is_hot = roll_race_weather(track_profile)
+
+    finishers = []
+    retired = []  # list of (driver, reason)
+
+    track_pace_w = track_profile.get("pace_weight", 1.0)
+    track_cons_w = track_profile.get("consistency_weight", 1.0)
+
+    # Race length factor (endurance = more failures)
+    race_distance_km = track_profile.get("race_distance_km", 250.0)
+    race_length_factor = race_distance_km / 250.0
+
+    reliability_mult = get_reliability_mult(time)
+    crash_mult = get_crash_mult(time)
+
+    event_grid = build_event_grid(state, time, race_name, track_profile)
+    grid_size = len(event_grid)
+    grid_risk_mult = 1.0 + max(0, grid_size - 12) * 0.01  # +1% per car above 12
+
+    for d in event_grid:
+        # Player shouldn't appear in AI-only race
+        if state.player_driver is d:
+            continue
+        if state.player_constructor and d.get("constructor") == state.player_constructor:
+            continue
+
+        ctor_speed, ctor_reliability = get_ai_car_stats(d["constructor"])
+        ctor_speed = max(1, ctor_speed)
+        reliability = ctor_reliability
+
+        # ---------- Performance roll (same vibe as your existing AI sim) ----------
+        base = d["pace"] * track_pace_w + d["consistency"] * 0.4 * track_cons_w
+        base += ctor_speed
+
+        cons_factor = max(0.0, min(d["consistency"] / 10.0, 0.95))
+        variance = (
+            random.uniform(-1, 1)
+            * (1 - cons_factor)
+            * base
+            * 0.25
+        )
+        performance = base + variance
+
+        # Wet pace effect
+        if is_wet:
+            wet_factor = d.get("wet_skill", 5) / 10.0
+            performance *= (0.9 + wet_factor * 0.3)
+
+        # Heat pace effect (small)
+        if is_hot:
+            heat_handling = (d.get("mechanical_sympathy", 5) + d.get("consistency", 5)) / 20.0
+            performance *= (0.97 + heat_handling * 0.06)
+
+        # ---------- DNF logic (ported from run_race, simplified for AI) ----------
+        mech = d.get("mechanical_sympathy", 5)
+        aggression = d.get("aggression", 5)
+        consistency = d.get("consistency", 5)
+        wet_skill = d.get("wet_skill", 5)
+
+        # Engine fail chance
+        engine_fail_chance = (11 - reliability) * 0.02 * reliability_mult
+        engine_fail_chance *= (1 + (5 - mech) * 0.05)
+        engine_fail_chance *= track_profile.get("engine_danger", 1.0)
+        engine_fail_chance *= race_length_factor
+
+        # Hot-day engine stress (AI heat tolerance assumed average=5)
+        if is_hot:
+            heat_intensity = track_profile.get("heat_intensity", 1.0)
+            engine_fail_chance *= heat_intensity
+
+        # Crash chance
+        base_crash_chance = (11 - consistency) * 0.012
+        base_crash_chance *= (1 + (aggression - 5) * 0.05)
+        base_crash_chance *= (1 + (5 - mech) * 0.03)
+
+        crash_chance = base_crash_chance * crash_mult
+        crash_chance *= track_profile.get("crash_danger", 1.0)
+
+        # Wet -> more crashes, better wet_skill reduces it
+        if is_wet:
+            wet_factor = wet_skill / 10.0
+            rain_crash_mult = 1.40 - wet_factor * 0.30
+            crash_chance *= rain_crash_mult
+
+        # Suspension affects crash risk (AI too)
+        sus = get_suspension_value_for_driver(state, d)
+        sus_importance = suspension_track_factor(track_profile)
+
+        crash_sus_mult = 1.08 - (sus - 5) * 0.02
+        crash_sus_mult = clamp(crash_sus_mult, 0.88, 1.15)
+        crash_sus_mult = 1.0 + (crash_sus_mult - 1.0) * sus_importance
+
+        crash_chance *= crash_sus_mult
+        crash_chance *= grid_risk_mult
+
+        # ---------- Resolve DNF ----------
+        if random.random() < engine_fail_chance:
+            retired.append((d, "engine"))
+            state.news.append(f"{d['name']} ({d['constructor']}) retired with engine failure.")
+
+            # --- simple breakdown so the news doesn't default to "general fatigue" ---
+            breakdown = []
+
+            reliability = max(1, ctor_reliability)
+            mech = d.get("mechanical_sympathy", 5)
+
+            reliability_mult = get_reliability_mult(time)
+            race_distance_km = track_profile.get("race_distance_km", 250.0)
+            race_length_factor = race_distance_km / 250.0
+
+            # Start from baseline
+            base_chance = (11 - reliability) * 0.02 * reliability_mult
+            running = base_chance
+
+            def add_factor(label, mult):
+                nonlocal running
+                extra = running * (mult - 1.0)
+                if extra > 0:
+                    breakdown.append((label, extra))
+                running *= mult
+
+            # Make reliability show up as a real contributor (lower reliability = more baseline risk)
+            # Weight isn't perfect science — it's just to stop "generic fatigue" lines.
+            reliability_weight = max(0.0, (11 - reliability) * 0.02 * reliability_mult)
+            breakdown.append(("car reliability", reliability_weight))
+
+            add_factor("driver mechanical sympathy", (1 + (5 - mech) * 0.05))
+            add_factor("track engine strain", track_profile.get("engine_danger", 1.0))
+            add_factor("race distance", race_length_factor)
+
+            if is_hot:
+                add_factor("heat intensity", track_profile.get("heat_intensity", 1.0))
+
+            breakdown.sort(key=lambda x: x[1], reverse=True)
+            if not breakdown:
+                breakdown = [("a run of bad luck", 1.0)]
+
+            add_engine_failure_explanation(
+                state, d, track_profile, is_hot, is_wet,
+                perspective="neutral",
+                breakdown=breakdown
+            )
+            continue
+
+        if random.random() < crash_chance:
+            retired.append((d, "crash"))
+            state.news.append(f"{d['name']} ({d['constructor']}) crashed out of the race.")
+            add_crash_explanation(state, d, track_profile, is_hot, is_wet, perspective="neutral")
+
+            # Check for injuries (player driver only)
+            if state.player_driver and d['name'] == state.player_driver['name']:
+                injury_roll = random.random()
+                if injury_roll < 0.05:  # 5% chance of career-ending injury
+                    state.player_driver_injury_severity = 3
+                    state.player_driver_injury_weeks_remaining = 0  # Immediate retirement
+                    state.news.append(f"TERRIBLE NEWS: {d['name']} has suffered a career-ending injury in the crash!")
+                    state.news.append(f"{d['name']} will never race again. Your team must find a new driver.")
+                    # Clear player driver
+                    state.player_driver = None
+                elif injury_roll < 0.20:  # 15% chance of serious injury (2-6 weeks)
+                    state.player_driver_injury_severity = 2
+                    weeks_out = random.randint(2, 6)
+                    state.player_driver_injury_weeks_remaining = weeks_out
+                    state.news.append(f"BAD NEWS: {d['name']} has suffered a serious injury in the crash!")
+                    state.news.append(f"{d['name']} will be unable to drive for {weeks_out} weeks.")
+                else:  # 80% chance of minor injury (1-2 weeks)
+                    state.player_driver_injury_severity = 1
+                    weeks_out = random.randint(1, 2)
+                    state.player_driver_injury_weeks_remaining = weeks_out
+                    state.news.append(f"{d['name']} has suffered a minor injury in the crash.")
+                    state.news.append(f"{d['name']} will be unable to drive for {weeks_out} week{'s' if weeks_out > 1 else ''}.")
+
+                state.player_driver_injured = state.player_driver_injury_weeks_remaining > 0
+
+            continue
+
+        # Survived -> classified finisher
+        finishers.append((d, performance))
+
+    if not finishers:
+        state.news.append(f"{race_name}: chaotic scenes — no cars reach the finish.")
+        record_race_result(state, time, season_week, race_name, is_wet, is_hot, finishers, retired)
+        state.completed_races.add(season_week)
+        return
+
+    # Sort finishers fastest to slowest
+    finishers.sort(key=lambda x: x[1], reverse=True)
+
+    # ------------------------------
+    # DEMO FINALE (AI-only): force fatal DNF so they cannot be classified
+    # ------------------------------
+    victim = maybe_trigger_demo_finale(state, time, race_name)
+    if victim:
+        vname = victim.get("name")
+
+        # If they were a finisher, pull them out of classification
+        before_n = len(finishers)
+        finishers = [(drv, perf) for (drv, perf) in finishers if drv.get("name") != vname]
+
+        # Ensure we log it as a retirement
+        retired.append((victim, "crash"))
+
+        if len(finishers) != before_n:
+            state.news.append(f"Classification update: {vname} is not classified after the incident.")
+
+        # Remove from global pool so future seasons are consistent
+        from gmr.data import drivers as global_drivers
+        if victim in global_drivers:
+            global_drivers.remove(victim)
+
+    # Fame/XP progression should apply to FINISHERS, not entrants
+    fame_mult = track_profile.get("fame_mult", 1.0)
+    xp_mult = track_profile.get("xp_mult", 1.0)
+
+    update_fame_after_race(
+        finishers,
+        fame_mult=fame_mult,
+        race_name=race_name,
+        season_week=season_week,
+        year=time.year
+    )
+
+    # Sponsor story event: driver promo at Fame 2+
+    from gmr.sponsorship import maybe_gallant_driver_promo
+    maybe_gallant_driver_promo(state, time)
+
+    update_driver_progress(state, finishers, time, xp_mult=xp_mult)
+
+    # Championship points (finishers only)
+    if CHAMPIONSHIP_ACTIVE:
+        for pos, (d, _) in enumerate(finishers):
+            if pos < len(POINTS_TABLE):
+                state.points[d["name"]] += POINTS_TABLE[pos]
+
+    # Headline with enhanced media flavor
+    winner = finishers[0][0]
+    if len(finishers) > 1:
+        runner_up = finishers[1][0]
+        headline = (
+            f"{race_name}: {winner['name']} wins for {winner['constructor']}, "
+            f"ahead of {runner_up['name']}."
+        )
+    else:
+        headline = f"{race_name}: {winner['name']} wins for {winner['constructor']}."
+    state.news.append(headline)
+
+    # Add atmospheric and media coverage based on race conditions
+    weather_descriptions = []
+    if is_wet:
+        weather_descriptions.extend([
+            "Dramatic wet-weather victory as rain made conditions treacherous.",
+            "Spectacular driving in the pouring rain - a true test of skill.",
+            "Rain-soaked triumph as drivers battled aquaplaning and poor visibility.",
+        ])
+    elif is_hot:
+        weather_descriptions.extend([
+            "Scorching conditions tested cars and drivers to their limits.",
+            "Heat haze shimmered over the track as temperatures soared.",
+            "Tires and engines pushed to the brink in the blazing heat.",
+        ])
+    else:
+        weather_descriptions.extend([
+            "Perfect racing conditions produced an exciting spectacle.",
+            "Clear skies set the stage for a thrilling motor race.",
+            "Sunshine and ideal weather delighted the crowd.",
+        ])
+    if weather_descriptions:
+        state.news.append(random.choice(weather_descriptions))
+
+    record_race_result(state, time, season_week, race_name, is_wet, is_hot, finishers, retired)
+
+
+def simulate_qualifying(state, race_name, time, track_profile):
+    """
+    Simulate a single qualifying session for this race.
+    """
+    results = []
+    grid_bonus = {}
+
+    is_wet_quali = random.random() < track_profile.get("wet_chance", WEATHER_WET_CHANCE)
+    track_pace_w = track_profile.get("pace_weight", 1.0)
+    track_cons_w = track_profile.get("consistency_weight", 1.0)
+
+    event_grid = build_event_grid(state, time, race_name, track_profile)
+
+    for d in event_grid:
+        # Track-specific pace weighting
+        base_pace = d["pace"] * track_pace_w
+        base_cons = d["consistency"] * 0.4 * track_cons_w
+
+        # Car performance for player vs AI
+        if d == state.player_driver:
+            car_speed = get_car_speed_for_track(state, track_profile)
+        else:
+            car_speed, _ = get_ai_car_stats(d["constructor"])
+
+        perf = base_pace + base_cons + car_speed
+
+        if is_wet_quali:
+            wet_factor = d.get("wet_skill", 5) / 10.0
+            perf *= (0.9 + wet_factor * 0.3)
+
+        # Small variance in qualifying (more swingy than race)
+        perf *= random.uniform(0.94, 1.06)
+
+        results.append((d, perf))
+
+    # Sort by performance
+    results.sort(key=lambda x: x[1], reverse=True)
+
+    # Grid bonuses are a tiny advantage for qualifying position
+    for i, (d, _perf) in enumerate(results):
+        grid_bonus[d["name"]] = 1.00 - (i * 0.003)
+
+    # Qualifying summary for news
+    if len(results) >= 2:
+        p1 = results[0][0]
+        p2 = results[1][0]
+        if p1["constructor"] == p2["constructor"]:
+            summary = (
+                f"{p1['name']} takes pole for {p1['constructor']} with teammate {p2['name']} alongside."
+            )
+        else:
+            summary = (
+                f"{p1['name']} takes pole for {p1['constructor']}, "
+                f"{p2['name']} joins on the front row."
+            )
+    elif len(results) == 1:
+        p1 = results[0][0]
+        summary = f"{p1['name']} takes pole for {p1['constructor']}"
+    else:
+        summary = "Qualifying completed."
+
+    session_weather = "wet" if is_wet_quali else "dry"
+    state.news.append(
+        f"Qualifying for {race_name} ({session_weather} session): {summary}"
+    )
+
+    # Results (Q1/Q2/Q3...) into news
+    state.news.append("Qualifying results:")
+    for pos, (d, _) in enumerate(results, start=1):
+        state.news.append(f"Q{pos}: {d['name']} ({d['constructor']})")
+
+    # Store quali results in state for overtake calculation
+    state.last_quali_results = results
+
+    return results, grid_bonus, is_wet_quali
+
+
+def run_race(state, race_name, time, season_week, grid_bonus, is_wet, is_hot):
+    state.news.append(f"=== {race_name} ===")
+
+    # How swingy race performance is (qualifying is higher)
+    variance_scale = 0.25
+
+    if state.player_driver:
+        eid = state.current_engine.get("unit_id") if state.current_engine else None
+        state.news.append(
+            f"DEBUG ENGINE: unit_id={eid}, wear={state.engine_wear:.0f}%, health={state.engine_health:.0f}%")
+
+    # PATCH : remember if we've ever completed Vallone GP
+    if race_name == "Vallone GP":
+        state.ever_completed_vallone = True
+
+    finishers = []
+    dnf_drivers = []
+
+    # Track why drivers retired this race (engine vs crash etc.)
+    retire_reasons = {}  # driver_name -> "engine" or "crash"
+
+    track_profile = tracks.get(
+        race_name,
+        {
+            "engine_danger": 1.0,
+            "crash_danger": 1.0,
+            "pace_weight": 1.0,
+            "consistency_weight": 1.0,
+            "wet_chance": WEATHER_WET_CHANCE,
+            "base_hot_chance": 0.2,
+            "heat_intensity": 1.0,
+            "race_distance_km": 250.0,
+        },
+    )
+
+    # Race length factor: how long this race is relative to a baseline 250 km
+    race_distance_km = track_profile.get("race_distance_km", 250.0)
+    race_length_factor = race_distance_km / 250.0
+
+    # ------------------------------
+    # WEATHER: WET vs DRY + HOT DAYS
+    # ------------------------------
+
+    if is_wet:
+        state.news.append("Rain falls over the circuit – conditions are WET.")
+    else:
+        if is_hot:
+            state.news.append("Baking heat grips the circuit – HOT and dry conditions.")
+        else:
+            state.news.append("Skies stay clear – conditions are DRY.")
+
+    event_grid = build_event_grid(state, time, race_name, track_profile)
+    grid_size = len(event_grid)
+    grid_risk_mult = 1.0 + max(0, grid_size - 12) * 0.01  # +1% per car above 12
+
+    stage_incidents = precompute_ai_stage_incidents(
+        event_grid,
+        state,
+        track_profile,
+        time,
+        is_wet,
+        is_hot,
+        grid_risk_mult,
+        race_length_factor,
+    )
+
+    # Quali results (stored in state from simulate_qualifying)
+    quali_results = getattr(state, "last_quali_results", [])
+    stage_overtakes = None
+
+    # 3-part race flow with player decisions/events
+    stage_mods = {
+        "performance_mult": 1.0,
+        "engine_fail_mult": 1.0,
+        "crash_mult": 1.0,
+    }
+    stage_output = []
+    player_in_grid = any(d == state.player_driver for d in event_grid)
+    if player_in_grid:
+        stage_mods, stage_output = run_player_race_stages(
+            state, race_name, is_wet, is_hot, track_profile,
+            stage_incidents=stage_incidents,
+            stage_overtakes=None,
+            defer_output=False
+        )
+        consume_tyre_set(state, race_name, "race start")
+
+    for d in event_grid:
+        # existing race logic unchanged
+
+        # --- Base driver performance with track bias ---
+        track_pace_w = track_profile.get("pace_weight", 1.0)
+        track_cons_w = track_profile.get("consistency_weight", 1.0)
+        weight_pace_importance = track_profile.get("weight_pace_importance", 1.0)
+        weight_crash_importance = track_profile.get("weight_crash_importance", 1.0)
+
+        base_pace = d["pace"] * track_pace_w
+        base_cons = d["consistency"] * 0.4 * track_cons_w
+
+        car_xp = float(d.get("car_xp", 0.0)) if d == state.player_driver else 0.0
+        car_xp = clamp(car_xp, 0.0, 10.0)
+
+        # Consistency controls how wild the variance is
+        consistency_factor = (d["consistency"] / 10) * track_cons_w
+        consistency_factor = max(0.0, min(consistency_factor, 0.95))
+
+        sus = get_suspension_value_for_driver(state, d)
+        sus_importance = suspension_track_factor(track_profile)
+
+        # 1–10 where 5 is baseline. Higher = less swing.
+        # var_mult range roughly ~1.22 (sus=1) to ~0.95 (sus=10) before track factor.
+        var_mult = 1.10 - (sus - 5) * 0.03
+        var_mult = clamp(var_mult, 0.85, 1.30)
+
+        # Track importance scales the effect away from 1.0
+        # If importance = 1.35, effect is stronger. If 0.85, effect is weaker.
+        var_mult = 1.0 + (var_mult - 1.0) * sus_importance
+
+        variance = (
+            random.uniform(-1, 1)
+            * (1 - consistency_factor)
+            * base_pace
+            * variance_scale
+            * track_pace_w
+            * var_mult
+        )
+
+        # Comfort reduces swinginess a bit (max ~20% reduction at 10.0)
+        variance *= (1.0 - car_xp * 0.02)
+
+        performance = base_pace + base_cons + variance
+
+        # Decide where the car performance comes from
+        if d == state.player_driver:
+            # Track-specific blend of engine speed vs acceleration
+            car_speed = get_car_speed_for_track(state, track_profile)
+
+            # Effective reliability lowered by long-term engine fatigue
+            health_factor = max(0.2, state.engine_health / 100.0)
+            car_reliability = state.car_reliability * health_factor
+        else:
+            car_speed, car_reliability = get_ai_car_stats(d["constructor"])
+
+        performance += car_speed
+
+        # Weather effect: in wet races, wet_skill matters for pace
+        if is_wet:
+            wet_factor = d["wet_skill"] / 10
+            performance *= (0.9 + wet_factor * 0.3)
+
+        # Heat effect
+        if is_hot:
+            heat_handling = (d["mechanical_sympathy"] + d["consistency"]) / 20.0
+            performance *= (0.97 + heat_handling * 0.06)
+
+        # Track-specific chassis weight effect on pace (player only)
+        if d == state.player_driver and state.current_chassis:
+            lightness = 11 - state.current_chassis["weight"]
+            pace_weight_effect = 1 + (lightness - 5) * 0.03 * weight_pace_importance
+            performance *= pace_weight_effect
+
+        # Apply race strategy
+        if d == state.player_driver:
+            mode = getattr(state, "risk_mode", "neutral")
+            if mode == "nurse":
+                # Slightly slower, safer
+                performance *= 0.97
+            elif mode == "attack":
+                # Slightly faster
+                performance *= 1.03
+        else:
+            # AI strategy
+            ai_mode = choose_ai_race_strategy(d, d["constructor"])
+            if ai_mode == "nurse":
+                performance *= 0.97
+            elif ai_mode == "attack":
+                performance *= 1.03
+            # neutral: no change
+
+        # Apply grid position bonus/penalty from qualifying
+        performance *= grid_bonus.get(d["name"], 1.0)
+
+        # Apply 3-stage race modifiers
+        if d == state.player_driver:
+            driver_stage_mods = stage_mods
+        else:
+            driver_stage_mods = run_ai_race_stages(d, is_wet, is_hot, race_length_factor, car_reliability)
+
+        performance *= driver_stage_mods["performance_mult"]
+
+        # --- DNF chances: engine vs crash ---
+        reliability_mult = get_reliability_mult(time)
+
+        # Engine failure from car reliability
+        reliability = car_reliability
+        engine_fail_chance = (11 - reliability) * 0.02 * reliability_mult
+
+        # Driver-side influence: mechanical sympathy
+        mech = d["mechanical_sympathy"]
+        engine_fail_chance *= (1 + (5 - mech) * 0.05)
+
+        # Track influence
+        engine_fail_chance *= track_profile["engine_danger"]
+
+        # Long-term engine condition: worn units fail more often
+        if d == state.player_driver:
+            wear_factor = max(0.2, min(state.engine_wear / 100.0, 1.0))  # clamp 0.2–1.0
+            # At 100%: *1.0; at 70%: ~1.3; at 50%: ~1.5; at 20%: ~1.8
+            engine_fail_chance *= (1 + (1 - wear_factor))
+
+        # Race length: endurance events expose engines longer
+        engine_fail_chance *= race_length_factor
+
+        # Engine heat tolerance: only meaningful on hot days
+        if d == state.player_driver and state.current_engine:
+            heat_tol = state.current_engine.get("heat_tolerance", 5)
+        else:
+            heat_tol = 5  # average for AI
+
+        if is_hot:
+            heat_intensity = track_profile.get("heat_intensity", 1.0)
+            engine_fail_chance *= heat_intensity
+            engine_fail_chance *= (1 + (5 - heat_tol) * 0.06)
+
+        # Crash chance from driver consistency
+        consistency = d["consistency"]
+        base_crash_chance = (11 - consistency) * 0.012
+        aggression = d["aggression"]
+        base_crash_chance *= (1 + (aggression - 5) * 0.05)
+        base_crash_chance *= (1 + (5 - mech) * 0.03)
+
+        # Player chassis effect on crash risk (light & worn chassis are twitchy)
+        if d == state.player_driver and state.current_chassis:
+            lightness = 11 - state.current_chassis["weight"]
+            base_crash_chance *= (1 + (lightness - 5) * 0.02 * weight_crash_importance)
+
+            # Extra risk when the chassis is tired
+            chassis_factor = (100.0 - state.chassis_health) / 100.0  # 0 at 100%, 1 at 0%
+            base_crash_chance *= (1 + chassis_factor * 0.5)  # up to +50% at 0% health
+
+        crash_mult = get_crash_mult(time)
+        crash_chance = base_crash_chance * crash_mult
+        crash_chance *= grid_risk_mult
+
+        # Track influence
+        crash_chance *= track_profile["crash_danger"]
+
+        # Long-term chassis condition: tired frames are scarier to crash in
+        if d == state.player_driver:
+            chassis_factor = max(0.2, min(state.chassis_wear / 100.0, 1.0))
+            # At 100%: *1.0; at 70%: ~1.33; at 50%: ~1.55; at 20%: ~1.88
+            crash_chance *= (1 + (1 - chassis_factor) * 1.1)
+
+        # Weather: wet -> more crashes, mitigated by wet_skill
+        if is_wet:
+            wet_factor = d["wet_skill"] / 10.0
+            rain_crash_mult = 1.40 - wet_factor * 0.30
+            crash_chance *= rain_crash_mult
+
+        sus = get_suspension_value_for_driver(state, d)
+        sus_importance = suspension_track_factor(track_profile)
+
+        # baseline around 5.
+        crash_sus_mult = 1.08 - (sus - 5) * 0.02
+        crash_sus_mult = clamp(crash_sus_mult, 0.88, 1.15)
+
+        crash_sus_mult = 1.0 + (crash_sus_mult - 1.0) * sus_importance
+        crash_chance *= crash_sus_mult
+
+        # Apply race strategy to failure risks (player only)
+        if d == state.player_driver:
+            mode = getattr(state, "risk_mode", "neutral")
+            if mode == "nurse":
+                engine_fail_chance *= 0.75
+                crash_chance *= 0.70
+            elif mode == "attack":
+                engine_fail_chance *= 1.25
+                crash_chance *= 1.30
+
+        # Apply 3-stage race modifiers to failure risks
+        engine_fail_chance *= driver_stage_mods["engine_fail_mult"]
+        crash_chance *= driver_stage_mods["crash_mult"]
+
+        # Comfort reduces crash risk slightly (max ~15% at 10.0)
+        crash_chance *= (1.0 - car_xp * 0.015)
+
+        # Force AI incidents to match stage updates
+        forced_incident = None
+        if d != state.player_driver:
+            forced_incident = stage_incidents.get(d.get("name"))
+            if forced_incident:
+                if forced_incident["type"] == "engine":
+                    engine_fail_chance = 1.0
+                    crash_chance = 0.0
+                elif forced_incident["type"] == "crash":
+                    crash_chance = 1.0
+                    engine_fail_chance = 0.0
+
+        # --- Build breakdown of what made engine risk high (for better news blurbs) ---
+        # Rank factors by how much extra failure probability they add above the base.
+        engine_fail_breakdown = []
+
+        base_chance = (11 - reliability) * 0.02 * reliability_mult  # reliability-only baseline
+        running = base_chance
+
+        def add_factor(label, mult):
+            nonlocal running
+            extra = running * (mult - 1.0)
+            if extra > 0:
+                engine_fail_breakdown.append((label, extra))
+            running *= mult
+
+        # Mechanical sympathy
+        mech_mult = (1 + (5 - mech) * 0.05)
+        add_factor("driver mechanical sympathy", mech_mult)
+
+        # Track strain
+        track_mult = track_profile.get("engine_danger", 1.0)
+        add_factor("track engine strain", track_mult)
+
+        # Wear + health (player only)
+        if d == state.player_driver:
+            wear_factor = max(0.2, min(state.engine_wear / 100.0, 1.0))
+            wear_mult = (1 + (1 - wear_factor))
+            add_factor("engine condition", wear_mult)
+
+            # Only mention long-term fatigue when it's actually meaningfully degraded
+            if state.engine_health < 75.0:
+                health_factor = max(0.2, state.engine_health / 100.0)
+                health_mult = 1 / health_factor
+                add_factor("long-term engine fatigue", health_mult)
+
+        # Race length
+        add_factor("race distance", race_length_factor)
+
+        # Heat (only if hot)
+        if is_hot:
+            heat_intensity = track_profile.get("heat_intensity", 1.0)
+            add_factor("heat intensity", heat_intensity)
+
+            heat_tol_mult = (1 + (5 - heat_tol) * 0.06)
+            add_factor("heat tolerance", heat_tol_mult)
+
+        # Sort by biggest extra-risk contributor
+        engine_fail_breakdown.sort(key=lambda x: x[1], reverse=True)
+
+        if not engine_fail_breakdown:
+            # If nothing actually increased risk, don't lie to the player.
+            # This covers "new engine, normal track, normal distance" failures.
+            if is_hot:
+                engine_fail_breakdown = [("heat stress", 1.0)]
+            elif track_profile.get("engine_danger", 1.0) > 1.02:
+                engine_fail_breakdown = [("track engine strain", 1.0)]
+            elif race_length_factor > 1.10:
+                engine_fail_breakdown = [("race distance", 1.0)]
+            else:
+                engine_fail_breakdown = [("an undetected component defect", 1.0)]
+
+        # ------------------------------
+        # PLAYER STAGE-BASED FAILURE CHECKS
+        # ------------------------------
+        player_stage_failure = None
+        player_stage_label = None
+        if d == state.player_driver:
+            stage_engine_chance = per_stage_chance(engine_fail_chance, stages=len(STAGE_LABELS))
+            stage_crash_chance = per_stage_chance(crash_chance, stages=len(STAGE_LABELS))
+
+            for stage_label in STAGE_LABELS:
+                if random.random() < stage_engine_chance:
+                    player_stage_failure = "engine"
+                    player_stage_label = stage_label
+                    break
+                if random.random() < stage_crash_chance:
+                    player_stage_failure = "crash"
+                    player_stage_label = stage_label
+                    break
+
+        # ------------------------------
+        # ENGINE FAILURE CHECK
+        # ------------------------------
+        if player_stage_failure == "engine" or (d != state.player_driver and random.random() < engine_fail_chance):
+            is_player = (d == state.player_driver)
+
+            if is_player and player_stage_label:
+                state.news.append(
+                    f"{player_stage_label}: {d['name']} ({d['constructor']}) slows with engine trouble."
+                )
+
+            if is_player and state.engine_wear < 50.0:
+                # Low-condition engine: chance it's completely destroyed
+                if random.random() < 0.4:
+                    state.news.append(
+                        f"{d['name']} ({d['constructor']}) suffers a huge engine blow-up – the unit is beyond saving."
+                    )
+                    state.engine_wear = 0.0
+
+                    # Hardware fatigue: future ceiling drops
+                    if state.engine_max_condition > 40.0:
+                        state.engine_max_condition -= 10.0
+
+                    # Scrap engine completely
+                    state.current_engine = None
+                    state.car_speed = 0
+                    state.car_reliability = 0
+
+                    state.news.append(
+                        "Your mechanics confirm the engine is scrap metal only – "
+                        "you'll need to buy or install a new unit before racing again."
+                    )
+
+                    retire_reasons[d["name"]] = "engine"
+
+                    add_engine_failure_explanation(
+                        state, d, track_profile, is_hot, is_wet,
+                        perspective="player",
+                        breakdown=engine_fail_breakdown
+                    )
+
+                else:
+                    state.news.append(
+                        f"{d['name']} ({d['constructor']}) retired with engine failure."
+                    )
+                    retire_reasons[d["name"]] = "engine"
+
+                    add_engine_failure_explanation(
+                        state, d, track_profile, is_hot, is_wet,
+                        perspective="player",
+                        breakdown=engine_fail_breakdown
+                    )
+
+            else:
+                # AI or healthier engines
+                state.news.append(
+                    f"{d['name']} ({d['constructor']}) retired with engine failure."
+                )
+                retire_reasons[d["name"]] = "engine"
+
+                add_engine_failure_explanation(
+                    state, d, track_profile, is_hot, is_wet,
+                    perspective="neutral",
+                    breakdown=engine_fail_breakdown
+                )
+            dnf_drivers.append(d)
+            continue  # stop processing this driver
+
+        # ------------------------------
+        # CRASH CHECK (ONLY IF ENGINE SURVIVED)
+        # ------------------------------
+
+        # --- Build breakdown of what made crash risk high (for better news blurbs) ---
+        crash_breakdown = []
+
+        base_c = (11 - consistency) * 0.012
+        running_c = base_c
+
+        def add_crash_factor(label, mult):
+            nonlocal running_c
+            extra = running_c * (mult - 1.0)
+            if extra > 0:
+                crash_breakdown.append((label, extra))
+            running_c *= mult
+
+        # Driver factors
+        add_crash_factor("driver aggression", (1 + (aggression - 5) * 0.05))
+        add_crash_factor("driver mechanical sympathy", (1 + (5 - mech) * 0.03))
+
+        # Track danger
+        add_crash_factor("track danger", track_profile.get("crash_danger", 1.0))
+
+        # Wet conditions (if wet)
+        if is_wet:
+            wet_factor = d.get("wet_skill", 5) / 10.0
+            rain_crash_mult = 1.40 - wet_factor * 0.30
+            add_crash_factor("wet conditions", rain_crash_mult)
+
+        # Suspension (only if it actually increased risk above baseline)
+        add_crash_factor("suspension compliance", crash_sus_mult)
+
+        # Player-only chassis condition multipliers
+        if d == state.player_driver and state.current_chassis:
+            lightness = 11 - state.current_chassis["weight"]
+            add_crash_factor("chassis lightness", (1 + (lightness - 5) * 0.02 * weight_crash_importance))
+
+            chassis_factor = max(0.2, min(state.chassis_wear / 100.0, 1.0))
+            add_crash_factor("chassis wear", (1 + (1 - chassis_factor) * 1.1))
+
+        crash_breakdown.sort(key=lambda x: x[1], reverse=True)
+        if not crash_breakdown:
+            crash_breakdown = [("a simple driving error", 1.0)]
+
+        if player_stage_failure == "crash" or (d != state.player_driver and random.random() < crash_chance):
+            if d == state.player_driver:
+                if player_stage_label:
+                    state.news.append(
+                        f"{player_stage_label}: {d['name']} ({d['constructor']}) loses the car in a critical moment."
+                    )
+                heavy_shunt = random.random() < 0.35
+
+                if heavy_shunt:
+                    state.news.append(
+                        f"{d['name']} ({d['constructor']}) has a heavy shunt and slams out of the race."
+                    )
+                else:
+                    state.news.append(
+                        f"{d['name']} ({d['constructor']}) loses it and crashes out of the race."
+                    )
+
+                add_crash_explanation(
+                    state, d, track_profile, is_hot, is_wet,
+                    perspective="player",
+                    breakdown=crash_breakdown
+                )
+
+                # Check for injuries
+                injury_roll = random.random()
+                if injury_roll < 0.05:  # 5% chance of career-ending injury
+                    state.player_driver_injury_severity = 3
+                    state.player_driver_injury_weeks_remaining = 0  # Immediate retirement
+                    state.news.append(f"TERRIBLE NEWS: {d['name']} has suffered a career-ending injury in the crash!")
+                    state.news.append(f"{d['name']} will never race again. Your team must find a new driver.")
+                    # Clear player driver
+                    state.player_driver = None
+                elif injury_roll < 0.20:  # 15% chance of serious injury (2-6 weeks)
+                    state.player_driver_injury_severity = 2
+                    weeks_out = random.randint(2, 6)
+                    state.player_driver_injury_weeks_remaining = weeks_out
+                    state.news.append(f"BAD NEWS: {d['name']} has suffered a serious injury in the crash!")
+                    state.news.append(f"{d['name']} will be unable to drive for {weeks_out} weeks.")
+                else:  # 80% chance of minor injury (1-2 weeks)
+                    state.player_driver_injury_severity = 1
+                    weeks_out = random.randint(1, 2)
+                    state.player_driver_injury_weeks_remaining = weeks_out
+                    state.news.append(f"{d['name']} has suffered a minor injury in the crash.")
+                    state.news.append(f"{d['name']} will be unable to drive for {weeks_out} week{'s' if weeks_out > 1 else ''}.")
+
+                state.player_driver_injured = state.player_driver_injury_weeks_remaining > 0
+
+                # Damage values
+                if heavy_shunt:
+                    chassis_hit = random.uniform(25.0, 45.0)
+                    engine_hit = random.uniform(15.0, 35.0)
+                else:
+                    chassis_hit = random.uniform(15.0, 30.0)
+                    engine_hit = random.uniform(5.0, 20.0)
+
+                old_chassis = state.chassis_wear
+                old_engine = state.engine_wear
+
+                state.chassis_wear = max(0.0, state.chassis_wear - chassis_hit)
+                state.engine_wear = max(0.0, state.engine_wear - engine_hit)
+
+                # Long-term fatigue
+                state.chassis_health = max(0.0, state.chassis_health - chassis_hit * 0.5)
+                state.engine_health = max(0.0, state.engine_health - engine_hit * 0.5)
+
+                # Chassis write-off chance
+                writeoff_chance = 0.15
+                if old_chassis < 50.0:
+                    writeoff_chance += 0.20
+                if heavy_shunt:
+                    writeoff_chance += 0.20
+
+                if random.random() < writeoff_chance:
+                    state.news.append(
+                        "The chassis is twisted beyond repair – the car is written off."
+                    )
+                    state.chassis_wear = 0.0
+                    if state.chassis_max_condition > 40.0:
+                        state.chassis_max_condition -= 10.0
+                    state.current_chassis = None
+                    state.car_speed = 0
+                else:
+                    state.news.append(
+                        f"Crash damage report: engine {old_engine:.0f}% → {state.engine_wear:.0f}%, "
+                        f"chassis {old_chassis:.0f}% → {state.chassis_wear:.0f}%."
+                    )
+
+                # Engine write-off from impact
+                engine_writeoff_chance = 0.0
+                if heavy_shunt:
+                    engine_writeoff_chance += 0.20
+                if old_engine < 40.0:
+                    engine_writeoff_chance += 0.20
+
+                if random.random() < engine_writeoff_chance:
+                    state.news.append(
+                        "Your mechanics report the engine block is cracked – the unit is beyond repair."
+                    )
+                    state.engine_wear = 0.0
+                    if state.engine_max_condition > 40.0:
+                        state.engine_max_condition -= 10.0
+                    state.current_engine = None
+                    state.car_speed = 0
+                    state.car_reliability = 0
+
+            else:
+                # AI crash
+                state.news.append(
+                    f"{d['name']} ({d['constructor']}) crashed out of the race."
+                )
+
+                add_crash_explanation(
+                    state, d, track_profile, is_hot, is_wet,
+                    perspective="neutral",
+                    breakdown=crash_breakdown
+                )
+
+            retire_reasons[d["name"]] = "crash"
+            dnf_drivers.append(d)
+            continue
+
+        # ------------------------------
+        # FINISHER (ONLY IF NO ENGINE FAIL + NO CRASH)
+        # ------------------------------
+        finishers.append((d, performance))
+
+    # Sort by performance
+    finishers.sort(key=lambda x: x[1], reverse=True)
+
+    # ------------------------------
+    # DEMO FINALE (player race): force fatal DNF so they cannot be classified
+    # ------------------------------
+    victim = maybe_trigger_demo_finale(state, time, race_name)
+    if victim:
+        vname = victim.get("name")
+
+        # Remove from classified finishers
+        finishers = [(drv, perf) for (drv, perf) in finishers if drv.get("name") != vname]
+
+        # Treat as crash retirement for prestige logic
+        retire_reasons[vname] = "crash"
+
+        # If victim is player's driver, they are a DNF and the contract ends brutally
+        if state.player_driver and state.player_driver.get("name") == vname:
+            dnf_drivers.append(state.player_driver)
+            state.demo_player_died = True  # just a flag for after-results cleanup
+
+        # Remove from global driver list for future seasons
+        from gmr.data import drivers as global_drivers
+        if victim in global_drivers:
+            global_drivers.remove(victim)
+
+    # Print accurate overtakes after final order is known
+    if player_in_grid:
+        stage_overtakes = compute_stage_overtakes_from_results(quali_results, finishers)
+        render_stage_overtakes_only(state, race_name, stage_overtakes)
+
+    # ------------------------------
+    # CAR COMFORT XP (player only)
+    # ------------------------------
+    if state.player_driver:
+        finished = any(d == state.player_driver for d, _ in finishers)
+        started = (state.player_driver in dnf_drivers) or finished
+
+        if started:
+            gain = 1.0 if finished else 0.35
+            state.player_driver["car_xp"] = round(
+                min(10.0, float(state.player_driver.get("car_xp", 0.0)) + gain), 2
+            )
+
+    # Track your driver's results with your team
+    player_finish_pos = None
+    if state.player_driver:
+        for pos, (d, _) in enumerate(finishers):
+            if d == state.player_driver:
+                player_finish_pos = pos
+                break
+
+        # They started the race, even if they DNF'd
+        state.races_entered_with_team += 1
+
+        if player_finish_pos is not None:
+            # Wins / podiums
+            if player_finish_pos == 0:
+                state.wins_with_team += 1
+                state.podiums_with_team += 1
+            elif player_finish_pos <= 2:
+                state.podiums_with_team += 1
+
+            # Championship points with your team
+            if player_finish_pos < len(POINTS_TABLE):
+                state.points_with_team += POINTS_TABLE[player_finish_pos]
+
+    fame_mult = track_profile.get("fame_mult", 1.0)
+    xp_mult = track_profile.get("xp_mult", 1.0)
+
+    update_fame_after_race(
+        finishers,
+        fame_mult=fame_mult,
+        race_name=race_name,
+        season_week=season_week,
+        year=time.year
+    )
+
+    # Sponsor story event: driver promo at Fame 2+
+    from gmr.sponsorship import maybe_gallant_driver_promo
+    maybe_gallant_driver_promo(state, time)
+
+    player_xp_gain = update_driver_progress(state, finishers, time, xp_mult=xp_mult)
+    player_xp_gain += grant_participation_xp_for_dnfs(state, dnf_drivers, time, xp_mult=xp_mult)
+
+    # ------------------------------
+    # CHAMPIONSHIP POINTS + PRIZE MONEY
+    # ------------------------------
+
+    # Award points ONLY if a championship exists
+    if CHAMPIONSHIP_ACTIVE:
+        for pos, (drv, _) in enumerate(finishers):
+            if pos < len(POINTS_TABLE):
+                state.points[drv["name"]] += POINTS_TABLE[pos]
+
+    # Pay prize money ONLY to the player's team (your cut of organiser prize)
+    if state.player_driver and player_finish_pos is not None:
+        raw_prize = get_prize_for_race_and_pos(race_name, player_finish_pos)  # player_finish_pos is 0-based
+        if raw_prize > 0:
+            prize_cut = int(raw_prize * CONSTRUCTOR_SHARE)
+
+            state.money += prize_cut
+            state.constructor_earnings += prize_cut
+            state.last_week_prize_income += prize_cut
+            state.last_week_income += prize_cut
+
+    # Appearance money (paid for taking the start; even if you DNF)
+    if state.player_driver:
+        pay_appearance_money(state, race_name)
+
+    # ------------------------------
+    # SPONSORSHIP PAYMENTS
+    # ------------------------------
+    if state.sponsor_active and state.player_driver:
+        # Appearance fee is paid as long as the car turned up for the race,
+        # even if it retires
+        state.sponsor_races_started += 1
+
+        mult = getattr(state, "sponsor_rate_multiplier", 1.0)
+
+        # Appearance money
+        appearance = int(60 * mult)
+        state.money += appearance
+        state.last_week_income += appearance
+        state.last_week_sponsor_income += appearance
+        state.constructor_earnings += appearance
+
+        # Points bonus ONLY if a championship exists
+        if CHAMPIONSHIP_ACTIVE and player_finish_pos is not None and player_finish_pos < len(POINTS_TABLE):
+            pts = POINTS_TABLE[player_finish_pos]
+            points_bonus = int(pts * 10 * mult)
+            state.money += points_bonus
+            state.last_week_income += points_bonus
+            state.last_week_sponsor_income += points_bonus
+            state.constructor_earnings += points_bonus
+
+        # Podium bonus (always valid, championship or not)
+        if player_finish_pos is not None and player_finish_pos <= 2:
+            state.sponsor_podiums += 1
+            podium_bonus = int(120 * mult)
+            state.money += podium_bonus
+            state.last_week_income += podium_bonus
+            state.last_week_sponsor_income += podium_bonus
+            state.constructor_earnings += podium_bonus
+
+        # Check sponsor goals
+        if not state.sponsor_goals_races_started and state.sponsor_races_started >= 3:
+            state.sponsor_goals_races_started = True
+            bonus = 500  # bonus for completing races started goal
+            state.money += bonus
+            state.last_week_income += bonus
+            state.last_week_sponsor_income += bonus
+            state.constructor_earnings += bonus
+            state.news.append(f"Sponsor bonus: Completed 3 races started goal! +£{bonus}")
+
+        if not state.sponsor_goals_podium and state.sponsor_podiums >= 1:
+            state.sponsor_goals_podium = True
+            bonus = 1000  # bonus for completing podium goal
+            state.money += bonus
+            state.last_week_income += bonus
+            state.last_week_sponsor_income += bonus
+            state.constructor_earnings += bonus
+            state.news.append(f"Sponsor bonus: Achieved first podium! +£{bonus}")
+
+    # Pay driver salary ONLY on race weeks (mercenary model)
+    if state.player_driver and state.driver_pay > 0:
+        state.money -= state.driver_pay
+        state.last_week_driver_pay = state.driver_pay
+        state.last_week_outgoings += state.driver_pay
+
+    # ------------------------------
+    # PATCH D: POST-RACE DEBRIEF (YOUR TEAM)
+    # ------------------------------
+    if state.player_driver:
+        # Store race-scoped numbers for UI/finances if needed
+        state.last_race_xp_gained = player_xp_gain
+        state.last_race_prize_gained = state.last_week_prize_income
+        state.last_race_sponsor_gained = state.last_week_sponsor_income
+        state.last_race_driver_pay = state.last_week_driver_pay
+
+        team_name = state.player_constructor or "Your team"
+        driver_name = state.player_driver["name"]
+
+        # How close to next stat tick?
+        current_xp_bank = state.player_driver.get("xp", 0.0)
+        xp_to_next = max(0.0, 5.0 - current_xp_bank)
+
+        state.news.append("")
+        state.news.append(f"--- Post-Race Debrief ({team_name}) ---")
+        state.news.append(f"Driver: {driver_name}")
+
+        # XP line
+        state.news.append(
+            f"Experience gained: +{player_xp_gain:.1f} XP "
+            f"(banked: {current_xp_bank:.1f}/5.0, {xp_to_next:.1f} to next improvement roll)"
+        )
+
+        # Money breakdown
+        prize = state.last_week_prize_income
+        sponsor = state.last_week_sponsor_income
+        appearance = getattr(state, "last_week_appearance_income", 0)
+        travel = getattr(state, "last_week_travel_cost", 0)
+        pay = state.last_week_driver_pay
+
+        state.news.append("Financial rundown:")
+        state.news.append(f"  Prize money (your cut): +£{prize}")
+        state.news.append(f"  Sponsor income: +£{sponsor}")
+        state.news.append(f"  Appearance / start money: +£{appearance}")
+        state.news.append(f"  Travel & logistics: -£{travel}")
+        state.news.append(f"  Driver pay: -£{pay}")
+
+        net = (prize + sponsor + appearance) - (travel + pay)
+        sign = "+" if net >= 0 else "-"
+        state.news.append(f"  Net from race weekend: {sign}£{abs(net)}")
+
+    # Race classification: show ALL classified finishers plus prize money
+    state.news.append("Race Results:")
+    for pos, (d, _) in enumerate(finishers):
+        place = pos + 1
+
+        prize = get_prize_for_race_and_pos(race_name, pos)
+
+        line = f"{place}. {d['name']} ({d['constructor']})"
+
+        # Only show points if the championship exists
+        if CHAMPIONSHIP_ACTIVE:
+            pts = POINTS_TABLE[pos] if pos < len(POINTS_TABLE) else 0
+            line += f" - {pts} pts"
+
+        if prize > 0:
+            line += f", Prize: £{prize}"
+            if d == state.player_driver:
+                ctor_share = int(prize * CONSTRUCTOR_SHARE)
+                line += f" (your cut: £{ctor_share})"
+
+        state.news.append(line)
+
+    # Save podium (top 3 finishers) for calendar display
+    top3 = []
+    for pos, (d, _) in enumerate(finishers[:3]):
+        top3.append((d["name"], d["constructor"]))
+    state.podiums[season_week] = top3
+    state.podiums_year = time.year
+
+    # --- Prestige gains for the player team (now fame-weighted) ---
+    if state.player_driver:
+        fame = state.player_driver.get("fame", 0)
+
+        # Step 1: base prestige change (same as before)
+        base_change = 0.0
+
+        if player_finish_pos is not None:
+            # Classified finish
+            if player_finish_pos == 0:
+                base_change = 3.0   # win
+            elif player_finish_pos == 1:
+                base_change = 2.0   # P2
+            elif player_finish_pos == 2:
+                base_change = 1.5   # P3
+            elif player_finish_pos < len(POINTS_TABLE):
+                base_change = 1.0   # points but no podium
+            else:
+                base_change = 0.3   # solid finish
+        else:
+            # DNF – split by cause
+            reason = retire_reasons.get(state.player_driver["name"])
+
+            if reason == "engine":
+                # Mechanical DNF – frustrating, but more “unlucky”
+                base_change = -0.2
+            elif reason == "crash":
+                # Crash DNF – more of a black mark
+                base_change = -0.8
+            else:
+                base_change = -0.5
+
+        # Step 2: fame multiplier (unchanged)
+        fame_capped = min(fame, 8)
+
+        if base_change > 0:
+            mult = 1.0 + fame_capped * 0.07
+        elif base_change < 0:
+            mult = 1.0 + fame_capped * 0.05
+        else:
+            mult = 1.0
+
+        change = base_change * mult
+
+        before = state.prestige
+        state.prestige = max(0.0, min(100.0, state.prestige + change))
+
+    # --- Post-race wear (player car only) ---
+    if state.player_driver:
+        base_engine_wear = 8.0    # typical GP
+        base_chassis_wear = 5.0   # chassis ages a bit slower
+
+        danger_factor_engine = track_profile.get("engine_danger", 1.0)
+        danger_factor_chassis = track_profile.get("crash_danger", 1.0)
+
+        engine_wear_loss = base_engine_wear * race_length_factor * danger_factor_engine
+        chassis_wear_loss = base_chassis_wear * race_length_factor * danger_factor_chassis
+
+        sus = int(state.current_chassis.get("suspension", 5)) if state.current_chassis else 5
+        sus_importance = suspension_track_factor(track_profile)
+
+        wear_sus_mult = 1.06 - (sus - 5) * 0.015
+        wear_sus_mult = clamp(wear_sus_mult, 0.90, 1.12)
+        wear_sus_mult = 1.0 + (wear_sus_mult - 1.0) * sus_importance
+
+        chassis_wear_loss *= wear_sus_mult
+
+        # Heat cooks engines; wet races are slightly gentler on the chassis
+        if is_hot:
+            engine_wear_loss *= 1.25
+        if is_wet:
+            chassis_wear_loss *= 0.9
+
+        # Small randomness so it isn't totally deterministic
+        engine_wear_loss *= random.uniform(0.8, 1.2)
+        chassis_wear_loss *= random.uniform(0.8, 1.2)
+
+        # Apply to condition, clamp at 0
+        old_engine = state.engine_wear
+        old_chassis = state.chassis_wear
+
+        state.engine_wear = max(0.0, state.engine_wear - engine_wear_loss)
+        state.chassis_wear = max(0.0, state.chassis_wear - chassis_wear_loss)
+
+        state.news.append(
+            f"Post-race inspection: engine {old_engine:.0f}% → {state.engine_wear:.0f}%, "
+            f"chassis {old_chassis:.0f}% → {state.chassis_wear:.0f}%."
+        )
+
+    # --- Long-term wear from this race (player car only) ---
+    if state.player_driver:
+        # Base wear from track + distance
+        engine_base_wear = 4.0 * race_length_factor * track_profile.get("engine_danger", 1.0)
+        chassis_base_wear = 3.0 * race_length_factor * track_profile.get("crash_danger", 1.0)
+
+        # Pace mode multiplies wear
+        pace = getattr(state, "risk_mode", "neutral")
+        risk_mult = getattr(state, "risk_multiplier", 1.0)
+
+        engine_wear = engine_base_wear * risk_mult
+        chassis_wear = chassis_base_wear * risk_mult
+
+        # Extra punishment if you actually finished the full distance
+        player_finished = any(d == state.player_driver for d, _ in finishers)
+        if player_finished:
+            engine_wear *= 1.2
+            chassis_wear *= 1.2
+        else:
+            # If you retired early, you spared some miles
+            engine_wear *= 0.8
+            chassis_wear *= 0.8
+
+        # Apply wear and clamp
+        state.engine_health = max(0.0, state.engine_health - engine_wear)
+        state.chassis_health = max(0.0, state.chassis_health - chassis_wear)
+
+    # Small news blurbs at thresholds – match what the player sees
+    if state.engine_wear < 40:
+        state.news.append(
+            "Your mechanics warn that the engine is getting very tired (<40% condition)."
+        )
+    elif state.engine_wear < 70:
+        state.news.append(
+            "The engine is showing its age – performance and reliability may start to drop."
+        )
+
+    if state.chassis_wear < 40:
+        state.news.append(
+            "Your chassis is badly fatigued – any big hit could write it off."
+        )
+    elif state.chassis_wear < 70:
+        state.news.append(
+            "Cracks and flex are appearing in the chassis – the ride is getting rough."
+        )
+
+    # If the scripted finale killed the player's driver, wipe the contract now.
+    if getattr(state, "demo_player_died", False):
+        state.player_driver = None
+        state.driver_pay = 0
+        state.driver_contract_races = 0
+        state.demo_player_died = False
+
+    retired = []
+    for d in dnf_drivers:
+        retired.append((d, retire_reasons.get(d.get("name"), "unknown")))
+
+    # ✅ Contract tick ONLY after the race is finished
+    started_race = (
+        (state.player_driver in dnf_drivers) or
+        any(d == state.player_driver for d, _ in finishers)
+    )
+
+    tick_driver_contract_after_race_end(state, time, started_race)
+
+    record_race_result(state, time, season_week, race_name, is_wet, is_hot, finishers, retired)
+
+    # ✅ CRITICAL: stop the race repeating
+    state.completed_races.add(season_week)
+
+    # ✅ If you use pending_race_week, clear it so the week doesn't re-trigger
+    if getattr(state, "pending_race_week", None) == season_week:
+        state.pending_race_week = None
+        state.completed_races.add(season_week)
