+++ main.py
@@ -30,6 +30,7 @@ from gmr.ui_garage import (
     show_garage,
     show_engine_shop,
     show_chassis_shop,
+    show_tyre_shop,
     manage_chassis_development,
     handle_repairs,
     handle_test_day,
@@ -335,7 +336,8 @@ def run_game():
                         print("\n=== Racecar Parts ===")
                         print("1. Engines")
                         print("2. Chassis")
-                        print("3. Back to Garage Menu")
+                        print("3. Tyres")
+                        print("4. Back to Garage Menu")
 
                         parts_choice = input("> ").strip()
 
@@ -344,6 +346,8 @@ def run_game():
                         elif parts_choice == "2":
                             show_chassis_shop(state)
                         elif parts_choice == "3":
+                            show_tyre_shop(state)
+                        elif parts_choice == "4":
                             break
                         else:
                             print("Invalid choice.")
